refactor panels components
