+++ src/components/Panels/Tabs/TabContext.tsx
@@ -1,13 +1,9 @@
 import React from 'react';
 
-import { PanelTab } from './PanelsTabs';
-
 interface TabContextProps {
   currentTab: string;
-  changeTab(tabName: string, newTab: PanelTab): void,
 }
 
 export const TabContext = React.createContext<TabContextProps>({
   currentTab: '',
-  changeTab: () => undefined,
 });
