+++ src/components/Panels/Tabs/index.ts
@@ -1,0 +1,3 @@
+export * from './NewTab';
+export * from './Tab';
+export * from './TabContext';
